pyproject.toml
[project]
name = "comfyui-hunyuanvideo-foley"
version = "1.0.0"
description = "ComfyUI wrapper nodes for HunyuanVideo-Foley: Generate audio from video + text prompts"
license = { text = "MIT" }
requires-python = ">=3.8"
classifiers = [
    "Operating System :: OS Independent",
    "Environment :: GPU :: NVIDIA CUDA",
]
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "torchaudio>=2.0.0",
    "numpy>=1.21.0",
    "scipy",
    "pillow",
    "av",
    "einops",
    "pyyaml",
    "omegaconf",
    "easydict",
    "loguru",
    "tqdm",
    "pandas",
    "pyarrow",
]

[project.urls]
Repository = "https://github.com/Jarcis-cy/ComfyUI-HunyuanVideoFoley"

[tool.comfy]
PublisherId = "Jarcis-cy"
DisplayName = "HunyuanVideo-Foley Audio Generator"
Icon = ""
